% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{amenities}
\alias{amenities}
\title{Function to recover the amenity distribution of working in each location.
Finds B's that make the probability of working in each sector in equation
(3.2) equal to the probabilities observed in the data.}
\usage{
amenities(theta1, kappa1, lambda_is_i, B_init, w_tr, reptau, tol, maxiter)
}
\arguments{
\item{theta1}{Float - Commuting elasticity.}

\item{kappa1}{Float - Parameter that governs the reallocation of workers
across sectors. It measures how easy it is for workers to substitute
jobs across sectors. In the case in which it tends to infity, the model
replicates the case in which there are no frictions in the labor market,
and workers are perfectly mobile across sectors. The case in which it
tends to one replicates the specific factor model in which workers are
only productive in one sector. Then, they do not switch jobs across
industries.}

\item{lambda_is_i}{NxS matrix - Proportion of residents in each location in
each sector. Rows add up to 1.}

\item{B_init}{Matrix - Initialization parameter for the amenity distribution
of work.}

\item{w_tr}{Matrix - Equilibrium wages}

\item{reptau}{NxN matrix - Commuting costs between all possible locations.}

\item{tol}{Float - Maximum tolerable error for estimating total labor.
Default tol=10^-6.}

\item{maxiter}{Integer - Maximum number of iterations for convergence.
Default maxiter=1000.}
}
\value{

}
\description{
Function to recover the amenity distribution of working in each location.
Finds B's that make the probability of working in each sector in equation
(3.2) equal to the probabilities observed in the data.
}
