% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{eq_quantities}
\alias{eq_quantities}
\title{Title}
\usage{
eq_quantities(
  alpha1,
  beta1,
  eta1,
  kappa1,
  sigma1,
  xi1,
  L_bar,
  H_bar,
  H_bar_rest,
  N,
  S,
  A_init,
  F,
  lambda_i,
  lambda_is_i,
  w,
  W_is,
  B,
  lambda_ijs_is,
  zeta_init = zeta_init,
  zeta,
  tol,
  maxiter,
  zeta_mult = c(),
  zeta_intervals = c()
)
}
\arguments{
\item{zeta_intervals}{}
}
\description{
Title
}
