% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{inversionModel_Eff}
\alias{inversionModel_Eff}
\title{Title}
\usage{
inversionModel_Eff(
  N,
  S,
  L_bar,
  H_bar,
  H_bar_rest,
  tau,
  lambda_i,
  lambda_is_i,
  L_j_data,
  zeta,
  z_init,
  tol = 10^-6,
  maxiter = 1000,
  alpha1 = 0.7,
  beta0 = 0.7,
  theta1 = 7,
  eta1 = 1.5,
  kappa1 = 2,
  sigma0 = 6,
  xi1 = 1.8,
  nu_init = 0.005,
  F = 1,
  nu_intervals = c(),
  nu_mult = c(),
  zeta_intervals = c(),
  zeta_mult = c()
)
}
\arguments{
\item{zeta_mult}{}
}
\value{

}
\description{
Title
}
