% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{labor_productivity}
\alias{labor_productivity}
\title{Function to compute equilibrium wages that make the model labor in every
location in every sector equal to the observed data. It finds the w's
such that equation (3.2) holds.}
\usage{
labor_productivity(
  N,
  S,
  w_init,
  theta1,
  tau,
  lambda_is_i,
  lambda_i,
  L_j_data,
  nu_init = 0.005,
  tol = 10^-6,
  maxiter = 1000,
  nu_mult = c(),
  nu_intervals = c()
)
}
\arguments{
\item{N}{Integer - Number of locations.}

\item{S}{Integer - Number of sectors in the economy.}

\item{w_init}{Wage initialization.}

\item{theta1}{Float - Commuting elasticity.}

\item{tau}{NxN matrix - Commuting costs between all possible locations.}

\item{lambda_is_i}{NxS matrix - Proportion of residents in each location in
each sector. Rows add up to 1.}

\item{lambda_i}{Nx1 matrix - Number of residents in each location.}

\item{L_j_data}{NxS matrix - Total amount of workers in each location in each
sector.}

\item{nu_init}{Float - Convergence parameter to update wages.
Default nu=0.005.}

\item{tol}{Float - Maximum tolerable error for estimating total labor.
Default tol=10^-6.}

\item{maxiter}{Integer - Maximum number of iterations for convergence.
Default maxiter=1000.}

\item{nu_mult}{Vector - Multipliers of nu.
Default is empty.}

\item{nu_intervals}{Vector - Intervals to use multipliers.
Default is empty.}
}
\value{
A list with equilibrium wages and probability of workers in each
location working in every other location.
}
\description{
Function to compute equilibrium wages that make the model labor in every
location in every sector equal to the observed data. It finds the w's
such that equation (3.2) holds.
}
