% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{living_amenities}
\alias{living_amenities}
\title{Function to estimate amenity parameters of locations where users live.}
\usage{
living_amenities(eta1, lambda_i, u_init, U_i, zeta, tol, maxiter)
}
\arguments{
\item{eta1}{Float - Parameter that governs the reallocation of workers across
locations in the city. This parameter measures how sensible are migration
flows within the city to changes in real income.}

\item{lambda_i}{Nx1 matrix - Probability of living in each location.}

\item{u_init}{Nx1 matrix - initialization matrix for amenity parameters.}

\item{U_i}{Matrix - Real income in each location.}

\item{zeta}{Float - Parameter to update amenity parameters in each iteration.}

\item{tol}{Float - Maximum tolerable error for estimating total labor.
Default tol=10^-6.}

\item{maxiter}{Integer - Maximum number of iterations for convergence.
Default maxiter=1000.}
}
\value{
Matrix with the amenity distribution of living in each location.
}
\description{
Function to estimate amenity parameters of locations where users live.
}
