% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_model.R
\name{living_amenities_simple}
\alias{living_amenities_simple}
\title{Function to estimate amenity parameters of locations where users live.}
\usage{
living_amenities_simple(theta, N, L_i, W_i, Q, K, alpha, t_ij, rho, eta)
}
\arguments{
\item{theta}{Float - Parameter that governs the reallocation of workers across
locations in the city. This parameter measures how sensible are migration
flows within the city to changes in real income.}

\item{L_i}{Nx1 matrix - Total residents.}

\item{W_i}{Nx1 matrix - Market access measure in each location.}

\item{Q}{Nx1 matrix - Floor space prices.}

\item{alpha}{Float - Para}

\item{t_ij}{NxN matrix - Travel times across locations.}

\item{rho}{Float - decay parameter for amenities.}

\item{eta}{Float - congestion force}
}
\value{
Matrix with the amenity distribution of living in each location.
}
\description{
Function to estimate amenity parameters of locations where users live.
}
