% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{number_firms}
\alias{number_firms}
\title{Computes number of firms}
\usage{
number_firms(N, beta1, beta_tilde_sector, w, L_j_eff, sigma1, q, F)
}
\arguments{
\item{N}{Int - Number of locations.}

\item{beta1}{Float - Cobb-Douglas parameter for the production function on
each sector.}

\item{beta_tilde_sector}{Float - (1-beta)/beta}

\item{w}{NxS - Wages in each location in each sector.}

\item{sigma1}{\itemize{
\item elasticity of subsitution across varieties within each sector.
}}

\item{q}{\itemize{
\item Commercial floorspace prices.
}}
}
\value{

}
\description{
Computes number of firms
}
