% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{price_location}
\alias{price_location}
\title{Computes prices charged by firms in each location to consumers in each
location.}
\usage{
price_location(N, sigma1, beta_factor, w, q, A)
}
\arguments{
\item{N}{Int - Number of locations.}

\item{sigma1}{\itemize{
\item elasticity of subsitution across varieties within each sector.
}}

\item{w}{NxS - Wages in each location in each sector.}

\item{q}{\itemize{
\item Commercial floorspace prices.
}}

\item{A}{}
}
\value{

}
\description{
Computes prices charged by firms in each location to consumers in each
location.
}
