% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{solveModel1_Eff}
\alias{solveModel1_Eff}
\title{Title}
\usage{
solveModel1_Eff(
  N,
  S,
  L_bar,
  H_bar,
  H_bar_rest,
  tau,
  A,
  u_eq,
  B,
  w_eq,
  lambda_i,
  lambda_is_i,
  zeta,
  tol = 10^-6,
  maxiter = 1000,
  endo_Lr = 1,
  alpha1 = 0.7,
  beta0 = 0.7,
  theta1 = 7,
  eta1 = 1.5,
  kappa1 = 2,
  xi1 = 1.8,
  sigma0 = 6,
  F = 1
)
}
\arguments{
\item{sigma0}{}
}
\value{

}
\description{
Title
}
