% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{av_income}
\alias{av_income}
\title{Computes average income in each location, which is the weighted average of
the income of the people living in the location.}
\usage{
av_income(lambda_is_i, B, W_is, kappa1, lambda_ijs_is, L_bar, w, lambda_i)
}
\arguments{
\item{lambda_is_i}{NxS matrix - Proportion of residents in each location in
each sector. Rows add up to 1.}

\item{B}{Amenity parameters of each location.}

\item{kappa1}{Float - Parameter that governs the reallocation of workers
across sectors. It measures how easy it is for workers to substitute
jobs across sectors. In the case in which it tends to infinity, the model
replicates the case in which there are no frictions in the labor market,
and workers are perfectly mobile across sectors. The case in which it
tends to one replicates the specific factor model in which workers are
only productive in one sector. Then, they do not switch jobs across
industries.}

\item{lambda_ijs_is}{NxNxS matrix - Probability of individuals in each
location of working in each location and each sector.}

\item{L_bar}{Float - Total number of people in the city.}

\item{w}{NxS - Wages in each location in each sector.}

\item{lambda_i}{Nx1 matrix - Number of residents in each location.}
}
\value{

}
\description{
Computes average income in each location, which is the weighted average of
the income of the people living in the location.
}
