% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{floorspace_supply}
\alias{floorspace_supply}
\title{Computes residential and comercial floorspace supply and equilibrium prices.}
\usage{
floorspace_supply(
  alpha1,
  H_bar_rest,
  H_bar,
  w,
  L_j_eff,
  y_bar,
  lambda_i,
  L_bar,
  beta1
)
}
\arguments{
\item{beta1}{}
}
\value{

}
\description{
Computes residential and comercial floorspace supply and equilibrium prices.
}
