% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_model.R
\name{inversionModel}
\alias{inversionModel}
\title{Function to invert model, so amenities, wages, productivities, and development density
are chosen to match model to data.}
\usage{
inversionModel(
  N,
  L_i,
  L_j,
  Q,
  K,
  t_ij,
  zeta,
  z_init,
  tol = 10^-10,
  maxiter = 1000,
  alpha,
  beta,
  theta,
  delta,
  rho,
  lambda,
  epsilon,
  mu,
  eta,
  nu_init = 0.005,
  nu_intervals = c(),
  nu_mult = c(),
  zeta_intervals = c(),
  zeta_mult = c()
)
}
\arguments{
\item{N}{Integer - Number of locations.}

\item{L_i}{Nx1 matrix - Number of residents in each location.}

\item{L_j}{Nx1 matrix - Number of workers in each location.}

\item{Q}{Nx1 matrix - Floorspace prices}

\item{K}{Nx1 matrix - Land area}

\item{t_ij}{NxN matrix - Travel times across all possible locations.}

\item{tol}{Int - tolerance factor}

\item{maxiter}{Integer - Maximum number of iterations for convergence.
Default maxiter=1000.}

\item{alpha}{Float - Utility parameter that determines preferences for
consumption.}

\item{beta}{Float - Output elasticity wrt labor}

\item{theta}{Float - Commuting elasticity and migration elasticity.}

\item{delta}{Float - Decay parameter agglomeration}

\item{rho}{Float - Decay parameter congestion}

\item{lambda}{Float - Agglomeration force}

\item{mu}{Float - Floorspace prod function: output elast wrt capital, 1-mu wrt land.}

\item{eta}{Float - Congestion force}

\item{zeta_intervals}{TODO}

\item{zeta_mult}{TODO}
}
\value{
Equilibrium values.
}
\description{
Function to invert model, so amenities, wages, productivities, and development density
are chosen to match model to data.
}
