% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{inversionModel_Eff}
\alias{inversionModel_Eff}
\title{Function to invert model, so amenities and wages are chosen to match model
to data.}
\usage{
inversionModel_Eff(
  N,
  S,
  L_bar,
  H_bar,
  H_bar_rest,
  tau,
  lambda_i,
  lambda_is_i,
  L_j_data,
  zeta,
  z_init,
  tol = 10^-6,
  maxiter = 1000,
  alpha1 = 0.7,
  beta0 = 0.7,
  theta1 = 7,
  eta1 = 1.5,
  kappa1 = 2,
  sigma0 = 6,
  xi1 = 1.8,
  nu_init = 0.005,
  F = 1,
  nu_intervals = c(),
  nu_mult = c(),
  zeta_intervals = c(),
  zeta_mult = c()
)
}
\arguments{
\item{N}{Integer - Number of locations.}

\item{S}{Integer - Number of sectors in the economy.}

\item{tau}{NxN matrix - Commuting costs between all possible locations.}

\item{lambda_i}{Nx1 matrix - Number of residents in each location.}

\item{lambda_is_i}{NxS matrix - Proportion of residents in each location in
each sector. Rows add up to 1.}

\item{L_j_data}{NxS matrix - Total amount of workers in each location in each
sector.}

\item{zeta}{Int - Convergence parameter for amenities}

\item{maxiter}{Integer - Maximum number of iterations for convergence.
Default maxiter=1000.}

\item{alpha1}{Float - Utility parameter that determines preferences for
consumption.}

\item{theta1}{Float - Commuting elasticity.}

\item{kappa1}{Float - Parameter that governs the reallocation of workers
across sectors. It measures how easy it is for workers to substitute
jobs across sectors. In the case in which it tends to infity, the model
replicates the case in which there are no frictions in the labor market,
and workers are perfectly mobile across sectors. The case in which it
tends to one replicates the specific factor model in which workers are
only productive in one sector. Then, they do not switch jobs across
industries.}

\item{zeta_mult}{}
}
\value{

}
\description{
Function to invert model, so amenities and wages are chosen to match model
to data.
}
