% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveModel.R
\name{solveModel}
\alias{solveModel}
\title{Function to solve counterfactuals.}
\usage{
solveModel(
  N,
  L_i,
  L_j,
  varphi,
  t_ij,
  K,
  a,
  b,
  w_eq,
  u_eq,
  Q_eq,
  ttheta_eq,
  alpha = 0.7,
  beta = 0.7,
  theta = 7,
  mu = 0.3,
  delta = 0.3585,
  lambda = 0.01,
  rho = 0.9094,
  eta = 0.1548,
  epsilon = 0.01,
  maxiter = 100,
  tol = 10^-10
)
}
\arguments{
\item{N}{Integer - Number of locations.}

\item{L_i}{Nx1 array - Number of residents in each location}

\item{L_j}{Nx1 array - Number of workers in each location}

\item{varphi}{Nx1 array - Density of development}

\item{t_ij}{NxN matrix - Travel times across locations}

\item{K}{Nx1 array - Land supply}

\item{a}{Nx1 array - Total Factor Productivity in each location}

\item{b}{Nx1 array - Vector of amenities in each location}

\item{w_eq}{Nx1 array - Initial vector of wages}

\item{u_eq}{Nx1 array - Initial vector of welfare}

\item{Q_eq}{Nx1 array - Initial price for floorspace}

\item{ttheta_eq}{Nx1 array - Share of floorspace used commercially}

\item{alpha}{Float - Exp. share in consumption, 1-alpha exp. share in housing}

\item{beta}{Float - Output elasticity with respect to labor}

\item{theta}{Float - Commuting and migration elasticity.}

\item{mu}{Float - Floorspace prod function: output elasticity wrt capital}

\item{delta}{Float - Decay parameter agglomeration force}

\item{lambda}{Float - agglomeration externality}

\item{rho}{Float - decay parameter for amenities}

\item{eta}{Float - amenity externality}

\item{epsilon}{Float - Parameter that transforms travel times to commuting costs}

\item{maxiter}{Integer - Maximum number of iterations for convergence.
Default maxiter=1000.}

\item{zeta}{Float - convergence parameter}
}
\value{
Counterfactual values.
}
\description{
Function to solve counterfactuals.
}
