% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{solveModel1_Eff}
\alias{solveModel1_Eff}
\title{Function to solve counterfactuals.}
\usage{
solveModel1_Eff(
  N,
  S,
  L_bar,
  H_bar,
  H_bar_rest,
  tau,
  A,
  u_eq,
  B,
  w_eq,
  lambda_i,
  lambda_is_i,
  zeta,
  tol = 10^-6,
  maxiter = 1000,
  endo_Lr = 1,
  alpha1 = 0.7,
  beta0 = 0.7,
  theta1 = 7,
  eta1 = 1.5,
  kappa1 = 2,
  xi1 = 1.8,
  sigma0 = 6,
  F = 1
)
}
\arguments{
\item{N}{Integer - Number of locations.}

\item{S}{Integer - Number of sectors in the economy.}

\item{L_bar}{TODO}

\item{H_bar}{TODO}

\item{H_bar_rest}{TODO}

\item{tau}{NxN matrix - Commuting costs between all possible locations.}

\item{A}{TODO}

\item{u_eq}{TODO}

\item{B}{TODO}

\item{w_eq}{TODO}

\item{lambda_i}{Nx1 matrix - Number of residents in each location.}

\item{lambda_is_i}{NxS matrix - Proportion of residents in each location in
each sector. Rows add up to 1.}

\item{zeta}{TODO}

\item{tol}{TODO}

\item{maxiter}{Integer - Maximum number of iterations for convergence.
Default maxiter=1000.}

\item{endo_Lr}{TODO}

\item{alpha1}{Float - Utility parameter that determines preferences for
consumption.}

\item{beta0}{TODO}

\item{theta1}{Float - Commuting elasticity.}

\item{eta1}{TODO}

\item{kappa1}{Float - Parameter that governs the reallocation of workers
across sectors. It measures how easy it is for workers to substitute
jobs across sectors. In the case in which it tends to infity, the model
replicates the case in which there are no frictions in the labor market,
and workers are perfectly mobile across sectors. The case in which it
tends to one replicates the specific factor model in which workers are
only productive in one sector. Then, they do not switch jobs across
industries.}

\item{xi1}{TODO}

\item{sigma0}{TODO}
}
\value{

}
\description{
Function to solve counterfactuals.
}
